+++ lib/rutabaga/feature.rb
@@ -4,6 +4,13 @@ require 'rspec'
 module Rutabaga
   module Feature
     def feature(feature_file = nil)
+      RSpec.deprecate(
+        "Calling `feature` from an `it` block",
+        :message => "Calling `feature` from an `it` block " \
+                    "is deprecated.\nIt should now be called directly in the " \
+                    "`describe` block."
+      )
+
       feature_file = Util.find_feature(feature_file || RSpec.current_example.description)
       example_group_class = self.class